requirements.txt
asgiref
backcall==0.2.0
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
decorator==5.0.9
dj-database-url==0.5.0
Django
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-mssql-backend==2.8.1
djangorestframework==3.12.4
drf-flex-fields==0.9.0
drf-nested-routers==0.93.3
fsspec==2021.6.0
gunicorn==20.1.0
hashids==1.3.1
idna==2.10
importlib-metadata==4.3.1
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.2
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2
pycparser==2.20
Pygments==2.9.0
pyhumps==3.0.2
PyMySQL==1.0.2
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
whitenoise==5.2.0
wincertstore==0.2
zipp==3.4.1
